\name{kernda-package}
\alias{kernda-package}
\docType{package}
\title{
Kernel Discriminant Analysis
}

\description{
Constrained ordinary least squares is performed. One constraint is that all beta coefficients
(including the constant) cannot be negative. They can be either 0 or strictly positive.
Another constraint is that the sum of the beta coefficients equals a constant.
References: Hansen, B.E. (2022). Econometrics, Princeton University Press.
}

\details{
\tabular{ll}{
Package: \tab kernda\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-06-29 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com}
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

%\references{
%Hansen, B. E. (2022). Econometrics, Princeton University Press.
%}





